Trying to level the plaing field for casting
It's having a really hard time modifying and/or adding behavior, and I
can only seem to get it to do one at a time (if at all). So,
copy/pasting `#slug` into `#as_json` so I can try to get it to *just*
delegate `#as_json` without the onerous task of also adding `#slug`.