Casting cannot implicitly override existing methods
The casting_test.rb seemed to assume that an object using casting could define existing
methods in attendant modules, but because it would rely on method_missing by declaring
delegate_missing_methods on the class, the existing methods would never be missing and
the module would not override anything.

The tests were altered so that differing call signatures could be reflected by the
particular style.